fix: degrade gracefully when packaged dependencies are missing (#39)
Move the real plugin to src/plugin.js and make the loader entry import it
behind a guard. When the installed package directory is incomplete (e.g. a
link: install whose node_modules was never populated), log one actionable
notice and export an inert plugin instead of failing the whole DSH plugin
tree with ERR_MODULE_NOT_FOUND at boot.
